# Conflicts:
#	src/shared/utility/CommunicationHeader.hpp

+++ src/shared/include/Obstacle.hpp
@@ -4,5 +4,5 @@ typedef struct obstacle_s {
     double x_pos;
     double y_pos;
 
-    obstacle_s(int x = 0, int y = 0) : x_pos(x), y_pos(y) {}
+    obstacle_s(double x = 0.0, double y = 0.0) : x_pos(x), y_pos(y) {}
 } obstacle_t;
